requirements.txt
pyyaml
requests
tqdm
sentence-transformers
torch
